관심 상품에 폴더 추가 및 폴더 전체 조회 구현

+++ src/main/java/com/sparta/myselectshop/repository/FolderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FolderRepository extends JpaRepository<Folder, Integer> {
+public interface FolderRepository extends JpaRepository<Folder, Long> {
 
     List<Folder> findAllByUserAndNameIn(User user, List<String> folderNames);
     // select * from folder where user_id = ? and name in (?, ?, ?...) --> in 안에는 folderNames 갯수만큼
